Some new stuff, should be more stable now.
git-svn-id: https://trac-hacks.org/svn@4197 7322e99d-02ea-0310-aa39-e9a107903beb

+++ pagetodocintegration/0.11/pagetodoc/pagetodoc.py
@@ -10,6 +10,8 @@ import os
 import glob
 import zipfile
 import base64
+import subprocess
+import popen2
 
 
 
@@ -22,13 +24,30 @@ class PageToDocPlugin(Component):
     tempdir = ''
     images = []
     imagesubdir = 'img/'
+    logsubdir = 'log/'
+    img_max_x = '0'
+    img_max_y = '0'
+    
+    verbose = False
 
     # IContentConverter methods
     def get_supported_conversions(self):
         yield ('zip', 'MS Word', 'zip', 'text/x-trac-wiki', 'application/zip', 7)
 
 
-    def convert_content(self, req, input_type, source, output_type):      
+    def convert_content(self, req, input_type, source, output_type):
+        
+        # get parameters from trac ini file
+        self.img_max_x = self.env.config.get('pagetodoc', 'img_max_x', self.img_max_x)
+        self.img_max_y = self.env.config.get('pagetodoc', 'img_max_y', self.img_max_y)
+        
+        # XSL-Transformation        
+        xsltfilepath = self.env.config.get('pagetodoc', 'xsltfile', '')
+        # TBD: Fehler ausgeben, wenn xsltfile nicht gelesen werden kann
+        # TBD: Parameter aus der trac.ini an zentraler Stelle auslesen
+        if False:
+            if xsltfilepath is '':
+                pass
         
         # maybe for later use
         #codepage = self.env.config.get('trac', 'charset', 'iso-8859-1')
@@ -46,8 +65,8 @@ class PageToDocPlugin(Component):
         
         # temporary files and folders
         self.tempdir = mkdtemp(prefix="page2doc")
-        htmlfilehandle, htmlfilepath = mkstemp(dir=self.tempdir)
-        wordfilehandle, wordfilepath = mkstemp(dir=self.tempdir)
+        htmlfilehandle, htmlfilepath = mkstemp(prefix='trac_', dir=self.tempdir)
+        wordfilehandle, wordfilepath = mkstemp(prefix='word_', dir=self.tempdir)
         zipfilepath = os.path.join(self.tempdir, os.path.basename(str(req.path_info) + '.zip'))
         
         # for debug: set all rights
@@ -69,31 +88,30 @@ class PageToDocPlugin(Component):
         os.close(htmlfilehandle)
         
         # clean up the HTML page using HTML Tidy
-        # maybe for later use
-        #tidy_input_enc = self.env.config.get('pagetodoc', 'input-encoding', 'utf8')
-        #tidy_output_enc = self.env.config.get('pagetodoc', 'output-encoding', 'latin1')   
-        #args = '-m -asxhtml --doctype omit --input-encoding %s --output-encoding %s' % (tidy_input_enc, tidy_output_enc)
         args = '-m -asxhtml -latin1 --doctype omit'
         cmd = 'tidy %s %s' % (args, htmlfilepath)
-        os.system(cmd)
+        self.execute_external_program(cmd)
         
-        # Workaround
-        self.xmlns_workaround(htmlfilepath)
+        # workaround namespace
+        self.perform_workarounds(htmlfilepath, 'html')
+    
+        cmd = 'xsltproc %s -o %s %s %s' % (('-v' if self.verbose else ''),  wordfilepath, xsltfilepath, htmlfilepath)
+        self.execute_external_program(cmd)
         
-        # XSL-Transformation        
-        xsltfilepath = self.env.config.get('pagetodoc', 'xsltfile', '')
-        cmd = 'xsltproc -o %s %s %s' % (wordfilepath, xsltfilepath, htmlfilepath)
-        os.system(cmd)
+        # workaround pre-tags
+        self.perform_workarounds(wordfilepath, 'pre')
+        
         
         # create a zip file and store all files into it      
         zipfilehandle = zipfile.ZipFile(zipfilepath, "w")
-        zipfilehandle.write(wordfilepath, os.path.basename(str(req.path_info) + '.htm'))
+        zipfilehandle.write(wordfilepath, os.path.basename(str(req.path_info) + '.htm'))       
         for image in self.images:
             zipfilehandle.write(image, self.imagesubdir + os.path.basename(image))     
         zipfilehandle.close()
         zip = open(zipfilepath, "rb").read()
         
         # delete temporary folders and files
+        self.remove_dir(os.path.join(self.tempdir, self.logsubdir))
         self.remove_dir(os.path.join(self.tempdir, self.imagesubdir))
         self.remove_dir(self.tempdir)
         
@@ -101,32 +119,92 @@ class PageToDocPlugin(Component):
         self.images = []
          
         return (zip, 'application/zip')
+    
+    
+    def execute_external_program(self, command):
+        logdir = os.path.join(self.tempdir, self.logsubdir)
+        self.create_dir(logdir)
+        
+        # Create output and error log files
+        outptr, outFile = mkstemp(dir=logdir)
+        errptr, errFile = mkstemp(dir=logdir)
+    
+        # Call the subprocess using convenience method
+        retval = subprocess.call(command, shell=True, stderr=errptr, stdout=outptr, close_fds=True)
+        
+        # read stdout and stderr
+        # its strange that all output goes to stderr instead of stdout, in both cases (error and no error)
+        # so always use stderr    
+        errptr = file(errFile, "r")
+        errData = errptr.read()
+        errptr.close()
+        
+        # log to trac.log
+        if self.verbose:
+            self.env.log.info('--------- EXTERNAL PROGRAM OUTPUT, command is ' + command)
+            self.env.log.info(errData)
+        
+        # not needed right now
+        #outptr = file(outFile, "r")
+        #outData = outptr.read()
+        #outptr.close()
+        
+        # Check the process exit code
+        if retval > 1:
+            raise Exception("Error executing command (return code = %s): %s" % (retval, errData))          
    
     # remove the xml namespace from the file
     # to be removed once I find out how to override this 
-    def xmlns_workaround(self, htmlfilepath):
+    def perform_workarounds(self, htmlfilepath, which=''):
          # Workaround: Entferne die Namespace-Angabe in der HTML-Datei
         htmlfilehandle = open(htmlfilepath, "r")
         html = htmlfilehandle.read()
         htmlfilehandle.close()
         
-        html = re.sub('(<html xmlns="http://www.w3.org/1999/xhtml">)', '<html>', html)
+        # replace namespace
+        if which is 'html':
+            html = re.sub('(<html xmlns="http://www.w3.org/1999/xhtml">)', '<html>', html)
+        
+        # remove line feeds in <pre>-tags    
+        if which is 'pre':
+            html = re.sub(r'<pre[^>]*>\n([^<]*)</pre>', self.remove_line_feeds, html)
         
         htmlfilehandle = open(htmlfilepath, "w")
         htmlfilehandle.write(html)
         htmlfilehandle.close()
+        
+    def remove_line_feeds(self, matchObj):
+        return '<pre>' + re.sub(r'\n', '<br />', matchObj.group(1)) + '<br /></pre>'
     
 
     def download_image(self, matchObj):
         imgdir = os.path.join(self.tempdir, self.imagesubdir)
         
-        if not os.path.isdir(imgdir):
-            os.mkdir(imgdir)
-            
-        fh, fn = mkstemp(dir = imgdir)
-        os.close(fh)
+        # create path to imagedir, if not existing
+        self.create_dir(imgdir)
         
+        # save image to disk
+        (filename, fileext) = os.path.splitext(os.path.basename(matchObj.group(1)))
+        # remove any trailing GET-Parameters from the file extension e.g. '.jpg?format=raw')
+        # fileext = fileext[:fileext.find('?')]
+        
+        # line above does not work due to encoding issues ('?' is something like '%3f' then)
+        # therefore just remove the file extention, and let ImageMagick detect the file type by itself
+        # this has been tested with ImageMagick 6.4.2 and PNG, JPG, TIFF and GIF files
+        fileext = ''
+        
+        # create temporary file 
+        fh, fn = mkstemp(prefix=filename, suffix=fileext, dir=imgdir)
+        os.close(fh)
         urlretrieve(matchObj.group(1), fn) 
+        
+        # resize images, if wanted, using ImageMagick
+        if int(self.img_max_x) > 0 and int(self.img_max_y) > 0:
+            args = "-resize '%sx%s>'" % (self.img_max_x, self.img_max_y)
+            cmd = 'convert %s %s %s' % (fn, args, fn)
+            self.execute_external_program(cmd)
+        
+        # add image to image list
         self.images.append(fn)
         
         return '<img src="%s"' % (self.imagesubdir + os.path.basename(fn))
@@ -139,13 +217,17 @@ class PageToDocPlugin(Component):
         for f in glob.glob(os.path.join(tempdir ,"*")):
             os.chmod(f, 0777)
     
-        
-    #removes a directory including all files    
+    # directory functions        
+    def create_dir(self, dir):
+        if not os.path.isdir(dir):
+            os.mkdir(dir)
+       
     def remove_dir(self, dir):
-        for f in glob.glob(os.path.join(dir, '*')):
-            os.unlink(f)
-        os.rmdir(dir)
-        
+        if os.path.isdir(dir):
+            for f in glob.glob(os.path.join(dir, '*')):
+                os.unlink(f)
+            os.rmdir(dir)
+
